requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
SQLAlchemy==2.0.30
pydantic==2.7.1
pydantic[email]==2.2.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.1